refactor(home): Changed the name of the home page file
